feat : Deduplicate header build logic
